+++ var/httpd/htdocs/js/Core.Agent.AppointmentCalendar.js
@@ -37,6 +37,7 @@ Core.Agent.AppointmentCalendar = (function (TargetNS) {
      */
     TargetNS.Init = function () {
         var CalendarConfig = Core.Config.Get('CalendarConfig'),
+            DefaultView,
             ResourceConfig,
             $CalendarObj = $('#calendar'),
             $DatepickerObj,
@@ -72,6 +73,14 @@ Core.Agent.AppointmentCalendar = (function (TargetNS) {
             };
         }
 
+        // Get default TimeLine from local storage if it is set.
+        if (typeof localStorage && localStorage.getItem('CalendarTimeLine') !== null) {
+            DefaultView = localStorage.getItem('CalendarTimeLine');
+        }
+        else {
+            DefaultView = Core.Config.Get('DefaultView');
+        }
+
         // Initialize calendar
         $CalendarObj.fullCalendar({
             header: {
@@ -87,7 +96,7 @@ Core.Agent.AppointmentCalendar = (function (TargetNS) {
                     }
                 }
             },
-            defaultView: Core.Config.Get('DefaultView'),
+            defaultView: DefaultView,
             allDayText: Core.Language.Translate('All-day'),
             isRTL: Core.Config.Get('IsRTLLanguage'),
             columnFormat: 'ddd, D MMM',
@@ -556,6 +565,11 @@ Core.Agent.AppointmentCalendar = (function (TargetNS) {
 
         $('.CalendarSwitch input[type="checkbox"]').each(function () {
             CalendarSwitchInit($(this));
+        });
+
+        // Set local storage for TimeLine overview.
+        $('.fc-left .fc-button-group button').on('click', function() {
+             localStorage.setItem('CalendarTimeLine', $(this).attr("class").match("fc-(month|agendaWeek|agendaDay|timelineMonth|timelineWeek|timelineDay)-button")[1]);
         });
 
         // Auto open appointment create screen.